provider/vault: Refactor client
In preparation for supporting a `vault_token` attribute on
`vault_secret` to override the provider's token on the client, I
realized that we couldn't call `client.SetToken` on the Vault client as
we were using it, since it was shared between all resources.

Refactored to have the resources depend on a `ClientProvider` interface
and have clients instantiated just-in-time so there's no sharing.

This probably breaks any HTTP connection reuse, but we can always circle back
around and make a performance pass as needed.